Allow building with GCC >= 13.3 on RISC-V
Building with GCC was marked as broken on RISC-V in
"46980605309e922d14f646c6705d184cb674c0f7". The patches fixing the issue
with atomic operations were included in the GCC 14 branch and backported
to 13.3.0.

Add the appropriate checks to allow building on RISC-V with the fixed
GCC versions.

Tested with GCC 13.3.0 on a VisionFive 2 board.

Change-Id: I9e4498640116b2b5fe73dd8e1822309444130998
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>