fix: atomic builtins defines for type support
Both UATOMIC_HAS_ATOMIC_BYTE and UATOMIC_HAS_ATOMIC_SHORT should only be
defined if the current architecture supports atomic operations on these
types.

The atomic builtins implementation maps these defines to the GCC/Clang
version that have a different semantics, where they are always defined
and the value determines if atomic operations are supported for the
type.

Only define UATOMIC_HAS_ATOMIC_XXX when the corresponding atomic builtin
define is set to '2' which indicates that the types is supported and
always lock-free.

Also reverse the order of the define checks since Clang and GCC both
define __GNUC__.

Change-Id: I781f5f2d81801c3690b478d90ffc546a8288fb6b
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>