Add builtin atomics size static asserts
Compiler builtin atomics support was introduced in 0.15 and is used for
some generic operations if the compiler has C11 support and for all
operations when configured with '--enable-compiler-atomic-builtins'.

The compiler builtins will accept any size and use a wrapper from
libatomic with locking for the sizes that the architecture has no atomic
operations for.

This is not something we want, this patch adds a static assert to detect
these operations and fail at build time for unsupported sizes.

Change-Id: I68fca42c57af042f79e935be7ae17d8ab9d0a9ff
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>