Move back CMM_LOAD/STORE_SHARED to volatile access
The CMM_{LOAD,STORE}_SHARED() API is intended to be usable on types that
are not necessarily supported by atomic operations on all platforms.

Move it back to a volatile access, all call sites both internal and in
the public headers that required ThreadSanitizer support we're moved to
uatomic_load/store in 4d197dd19c44 ("Tree-wide: Rename to
uatomic_load/uatomic_store").

Change-Id: I27a8680c7d7891cd118c5756c7e91462267b799b
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>