Use uatomic_load CMM_RELAXED in URCU_DEREFERENCE_USE_VOLATILE
uatomic_load() with relaxed mo implies a volatile access, which makes it
a good candidate to implement rcu_dereference() and replace the
CMM_LOAD_SHARED when URCU_DEREFERENCE_USE_VOLATILE is defined.

Although it does not track happens-before (due to lack of consume mo),
this has the advantage to mark the memory read as "atomic" from a
threadsanitizer perspective.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Id7356ea4a56bd81be77b1afd4d3b898905187405