Add cmm_annotate_mem_acquire() to URCU_DEREFERENCE_USE_VOLATILE rcu_dereference
When URCU_DEREFERENCE_USE_VOLATILE is defined, rcu_dereference uses a
volatile relaxed atomic load to fetch the RCU pointer. It relies on the
address dependency to ensure proper ordering of dependent loads. Inform
TSAN of this dependency by issuing an explicit __tsan_acquire() on the
address loaded from, which pairs with the release mo in
rcu_assign_pointer, rcu_set_pointer, and rcu_xchg_pointer.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I88714605d9e37ca9c5ccd220246fe10c4379dfd6