fix: urcu assert fallback for pre-C11 builds
It was discovered that the trick we used to implement static assertions
when building without C11 support would silently ignore some non
constant expression predicates. For example, the builtin
__atomic_always_lock_free() which doesn't resolve to a constant
expression on GCC < 5.1 fails as expected with _Static_assert() but is
silently ignored with the fallback.

Replace it by a new implementation based on defining a struct containing
a bitfield with a size determined by the integer value of the predicate.
When the predicate is false (0) the compilation fails because a zero
sized bitfield is invalid.

Change-Id: I4c4a7364ecb548801ae1c44fc9a117316a76707f
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>