fix: m4: ensure newline after the no-op command
The no-op builtin shell command ':' is prepended to the second argument
of many AS_IF calls to ensure there is at least one command in the
'else' branch of the condition. A 'dnl' m4 macro was added after the ':'
to remove the newline from the output of the macro which results in the
first line of the macro argument being passed as an argument to ':'
which is silently ignored.

Change-Id: I0bf0924b157e8d899bb9022577a9223a5c3c61cf
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>